Revert "moved the policy file from caravan-wallets to  caravan/multisig (#261)"
This reverts commit ff187d6ddefd23bfc3a5a5eab872223dadcdf04f.
