+++ apps/coordinator/src/components/Wallet/RegisterWallet.jsx
@@ -13,7 +13,7 @@ import {
 } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import { ExpandMoreOutlined } from "@mui/icons-material";
-import { MultisigWalletPolicy } from "@caravan/multisig";
+import { MultisigWalletPolicy } from "@caravan/wallets";
 
 import { getWalletConfig } from "../../selectors/wallet";
 import PolicyRegistrationTable from "../RegisterWallet/PolicyRegistrationsTable";
