+++ packages/caravan-wallets/package.json
@@ -43,6 +43,7 @@
     "@babel/preset-typescript": "^7.18.6",
     "@caravan/build-plugins": "*",
     "@caravan/eslint-config": "*",
+    "@caravan/multisig": "*",
     "@caravan/typescript-config": "*",
     "@inrupt/jest-jsdom-polyfills": "^3.2.1",
     "@types/jest": "^29.5.12",
@@ -80,7 +81,6 @@
     "@babel/polyfill": "^7.7.0",
     "@babel/runtime": "^7.23.9",
     "@caravan/bitcoin": "*",
-    "@caravan/multisig": "*",
     "@caravan/psbt": "*",
     "@ledgerhq/hw-app-btc": "^5.34.1",
     "@ledgerhq/hw-transport-u2f": "^5.34.0",
