+++ packages/caravan-wallets/src/coldcard.ts
@@ -27,8 +27,9 @@ import {
   INFO,
   ERROR,
 } from "./interaction";
-import { P2SH, P2SH_P2WSH, P2WSH, BitcoinNetwork } from "@caravan/bitcoin";
-import { WalletConfigKeyDerivation } from "@caravan/multisig";
+import { P2SH, P2SH_P2WSH, P2WSH } from "@caravan/bitcoin";
+import { BitcoinNetwork } from "@caravan/bitcoin";
+import { WalletConfigKeyDerivation } from "./types";
 import {
   convertLegacyInput,
   convertLegacyOutput,
