+++ packages/caravan-wallets/src/fixtures.ts
@@ -1,5 +1,4 @@
 import { TEST_FIXTURES, Network } from "@caravan/bitcoin";
-
 import { KeyOrigin, MultisigWalletPolicy } from "./policy";
 
 const paths = ["m/48'/1'/100'/1'", "m/48'/1'/100'/2'"];
