+++ packages/multisig/src/index.ts
@@ -1,3 +1,2 @@
-export * from "./types";
-export * from "./utils";
-export * from "./policy";
+export * from "./types"
+export * from "./utils"
