+++ packages/multisig/src/types.ts
@@ -1,5 +1,9 @@
 import { MultisigAddressType, BitcoinNetwork } from "@caravan/bitcoin";
 
+export interface DeviceError extends Error {
+  message: string;
+}
+
 // really should be interchangeable with KeyDerivations
 // but unfortunately there are inconsistent property names
 // in different contexts
