+++ packages/multisig/tsup.config.js
@@ -1,6 +1,6 @@
-import { nodeModulesPolyfillPlugin } from "esbuild-plugins-node-modules-polyfill";
-import { defineConfig } from "tsup";
+import { defineConfig } from 'tsup';
+import { polyfillNode } from "esbuild-plugin-polyfill-node";
 
 export default defineConfig({
-  esbuildPlugins: [nodeModulesPolyfillPlugin()],
+  esbuildPlugins: [polyfillNode()],
 });
